Refactored algorithms and exceptions
