+++ geoalt_exceptions/exceptions.py
